fixed fucking hibernate error

+++ src/main/java/com/server/tick/ResourceTick.java
@@ -32,7 +32,7 @@ public class ResourceTick implements Runnable {
         for (int i = 0; i < towns.size(); i++) {
             Town town = towns.get(i);
             OilBuilding building = (OilBuilding) buildingHandler.getResourceBuilding(town.getId(), ResourceType.OIL.getCode());
-            int production = building.getResourceProduction();
+            double production = building.getResourceProduction();
             //town.addResource(building.);
             //building.getResourceProduction();
         }
